+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Bug fixes
   * Allow you to change your password with `mix hex.user update`
   * Correctly display dependencies in `mix hex.info PACKAGE VERSION`
+  * Verify peer certificates when fetching tarball
 
 ## v0.5.0 (2014-09-19)